Task 150: remove redundant attributesvalues
