+++ src/main/java/taiga/model/query/customattributes/UserStoryCustomAttributesValues.java
@@ -2,43 +2,43 @@ package taiga.model.query.customattributes;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+import java.util.HashMap;
 
-    public class UserStoryCustomAttributesValues {
+public class UserStoryCustomAttributesValues {
 
     @SerializedName("attributes_values")
     @Expose
-    private AttributesValues attributesValues;
-
-    @SerializedName("user_story")
-    @Expose
-    private Integer userStory;
+    private HashMap<String, String> attributesValues;
 
     @SerializedName("version")
     @Expose
-    private Integer version;
+    private int version;
 
-    public AttributesValues getAttributesValues() {
-    return attributesValues;
+    @SerializedName("user_story")
+    @Expose
+    private int userStory;
+
+    public HashMap<String, String> getAttributesValues() {
+        return attributesValues;
     }
 
-    public void setAttributesValues(AttributesValues attributesValues) {
-    this.attributesValues = attributesValues;
+    public void setAttributesValues(HashMap<String, String> attributesValues) {
+        this.attributesValues = attributesValues;
     }
 
-    public Integer getUserStory() {
-    return userStory;
-    }
-
-    public void setUserStory(Integer userStory) {
-    this.userStory = userStory;
-    }
-
-    public Integer getVersion() {
-    return version;
+    public int getVersion() {
+        return version;
     }
 
     public void setVersion(Integer version) {
-    this.version = version;
+        this.version = version;
     }
 
+    public int getUserStory() {
+        return userStory;
+    }
+
+    public void setUserStory(Integer userStory) {
+        this.userStory = userStory;
+    }
 }
